chore(makeStyles): refactor useRenderer()
