+++ packages/make-styles/etc/make-styles.api.md
@@ -16,8 +16,8 @@ export function createCSSVariablesProxy(prefix?: string): unknown;
 
 // Warning: (ae-forgotten-export) The symbol "MakeStylesDOMRenderer" needs to be exported by the entry point index.d.ts
 //
-// @public (undocumented)
-export function createDOMRenderer(target: Document | undefined): MakeStylesDOMRenderer;
+// @public
+export function createDOMRenderer(target?: Document | undefined): MakeStylesDOMRenderer;
 
 // Warning: (ae-internal-missing-underscore) The name "DEFINITION_LOOKUP_TABLE" should be prefixed with an underscore because the declaration is marked as @internal
 //
