+++ packages/make-styles/src/makeStaticStyles.test.ts
@@ -1,18 +1,16 @@
-import { createDOMRenderer, MakeStylesDOMRenderer, resetDOMRenderer } from './renderer/createDOMRenderer';
-import { makeStaticStyles } from './makeStaticStyles';
+import { createDOMRenderer } from './renderer/createDOMRenderer';
 import { makeStylesRendererSerializer } from './utils/test/snapshotSerializer';
+import { makeStaticStyles } from './makeStaticStyles';
 import { makeStyles } from './makeStyles';
+import { MakeStylesRenderer } from './types';
 
 expect.addSnapshotSerializer(makeStylesRendererSerializer);
 
 describe('makeStaticStyles', () => {
-  let renderer: MakeStylesDOMRenderer;
+  let renderer: MakeStylesRenderer;
+
   beforeEach(() => {
     renderer = createDOMRenderer(document);
-  });
-
-  afterEach(() => {
-    resetDOMRenderer();
   });
 
   it('handles static styles', () => {
