+++ packages/react-make-styles/src/__styles.ts
@@ -1,7 +1,7 @@
 import { __styles as vanillaStyles, ResolvedStylesBySlots } from '@fluentui/make-styles';
 import { useFluent } from '@fluentui/react-shared-contexts';
 
-import { useRenderer } from './useRenderer';
+import { useRenderer } from './RendererContext';
 
 /**
  * A version of makeStyles() that accepts build output as an input and skips all runtime transforms.
@@ -13,8 +13,8 @@ export function __styles<Slots extends string>(resolvedStyles: ResolvedStylesByS
   const getStyles = vanillaStyles(resolvedStyles);
 
   return function useClasses(): Record<Slots, string> {
-    const { dir, targetDocument } = useFluent();
-    const renderer = useRenderer(targetDocument);
+    const { dir } = useFluent();
+    const renderer = useRenderer();
 
     return getStyles({ dir, renderer });
   };
