+++ packages/react-make-styles/src/index.ts
@@ -3,6 +3,8 @@ export { mergeClasses, createDOMRenderer } from '@fluentui/make-styles';
 export { makeStyles } from './makeStyles';
 export { makeStaticStyles } from './makeStaticStyles';
 
+export * from './RendererContext';
+
 // Private exports, are used by build time transforms
 export { __styles } from './__styles';
