+++ packages/react-make-styles/src/makeStaticStyles.ts
@@ -3,9 +3,8 @@ import {
   makeStaticStyles as vanillaMakeStaticStyles,
   MakeStaticStylesOptions,
 } from '@fluentui/make-styles';
-import { useFluent } from '@fluentui/react-shared-contexts';
 
-import { useRenderer } from './useRenderer';
+import { useRenderer } from './RendererContext';
 
 export function makeStaticStyles<Selectors>(styles: MakeStaticStyles | MakeStaticStyles[]) {
   const getStyles = vanillaMakeStaticStyles(styles);
@@ -16,9 +15,7 @@ export function makeStaticStyles<Selectors>(styles: MakeStaticStyles | MakeStati
   }
 
   return function useStaticStyles(): void {
-    const { targetDocument } = useFluent();
-
-    const renderer = useRenderer(targetDocument);
+    const renderer = useRenderer();
     const options: MakeStaticStylesOptions = { renderer };
 
     return getStyles(options);
