+++ packages/react-make-styles/src/makeStyles.ts
@@ -2,15 +2,15 @@ import { makeStyles as vanillaMakeStyles, MakeStylesOptions, MakeStylesStyleRule
 import { useFluent } from '@fluentui/react-shared-contexts';
 import { Theme } from '@fluentui/react-theme';
 
-import { useRenderer } from './useRenderer';
+import { useRenderer } from './RendererContext';
 
 export function makeStyles<Slots extends string>(stylesBySlots: Record<Slots, MakeStylesStyleRule<Theme>>) {
   const getStyles = vanillaMakeStyles(stylesBySlots);
 
   return function useClasses(): Record<Slots, string> {
-    const { dir, targetDocument } = useFluent();
+    const { dir } = useFluent();
 
-    const renderer = useRenderer(targetDocument);
+    const renderer = useRenderer();
     const options: MakeStylesOptions = {
       dir,
       renderer,
